Reorganizing library
