+++ .devcontainer/devcontainer.json
@@ -39,6 +39,8 @@
 			},
 			// Add the IDs of extensions you want installed when the container is created.
 			"extensions": [
+				// "usernamehw.errorlens",
+				"streetsidesoftware.code-spell-checker",
 				"GitHub.copilot",
 				"GitHub.copilot-chat",
 				"ms-vscode.cpptools-extension-pack",
