+++ dsg_lib/common_functions/folder_functions.py
