+++ dsg_lib/common_functions/patterns.py
