+++ examples/fastapi_example.py
@@ -13,6 +13,8 @@ from sqlalchemy import Column, ForeignKey, Select, String
 from sqlalchemy.orm import relationship
 from tqdm import tqdm
 
+from dsg_lib import logging_config
+
 # from dsg_lib import (
 #     async_database,
 #     base_schema,
@@ -21,9 +23,6 @@ from tqdm import tqdm
 #     logging_config,
 # )
 
-from dsg_lib.fastapi_functions import http_codes
-
-from dsg_lib import logging_config
 
 logging_config.config_log(
     logging_level="INFO", log_serializer=False, log_name="log.log"
