+++ xtests/test_common_functions/test_file_functions/test_save_text.py
@@ -4,7 +4,7 @@ import unittest
 from tempfile import TemporaryDirectory
 from unittest.mock import patch
 
-from dsg_lib.file_functions import save_text
+from src.dsg_lib.common_functions.file_functions import save_text
 
 
 class SaveTextTestCase(unittest.TestCase):
